Move query collection to a separate file
